refactor more stuff into service/core and service/start_node
