+++ cumulus/polkadot-parachain/src/service/start_nodes/mod.rs
@@ -2,3 +2,6 @@ pub mod asset_hub_lookahead;
 pub mod rococo_contracts;
 pub mod asset_hub;
 pub mod shell;
+pub mod basic_lookahead;
+pub mod generic_aura_lookahead;
+pub mod rococo_parachain;
